0211/app1/pyproject.toml
[project]
name = "app1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.7",
    "mariadb>=1.1.14",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
]
